task_1/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "binary_tree_pkg"
version = "0.1.0"
description = "A binary tree implementation with YAML"
dependencies = [
    "PyYAML",
]

[project.scripts]
# This allows you to run the test from command line if needed
